sw_apps: pmufw: Optimize PmSystemRequirement structure
Capabilities (1-hot) does require more than u8 size.
So use u8 for storing capabilities in PmSystemRequirement
data structure to reduce the memory requirement.

This reduces around 128 bytes of memory.


Acked-for-series: Mounika Akula <mounika.akula@xilinx.com>

+++ lib/sw_apps/zynqmp_pmufw/src/pm_system.c
@@ -68,8 +68,8 @@ typedef struct {
  */
 typedef struct PmSystemRequirement {
 	PmSlave* const slave;
-	u32 caps;
-	u32 posCaps;
+	u8 caps;
+	u8 posCaps;
 } PmSystemRequirement;
 
 #ifdef ENABLE_POS
